HMW-799 Extracted toggle table and date slider into component
